util/time: move raw_now to mod time
