+++ src/util/mod.rs
@@ -38,7 +38,6 @@ pub mod buf;
 pub mod transport;
 pub mod file;
 pub mod file_log;
-pub mod clocktime;
 pub mod metrics;
 pub mod threadpool;
 pub mod collections;
